app/requirements.txt
Flask
psycopg2-binary
fuzzywuzzy
Flask-SQLAlchemy
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp
opentelemetry-instrumentation-psycopg2
python-dotenv